+++ src/bl4ckscor3/bot/bl4ckb0t/core/Listener.java
@@ -40,6 +40,7 @@ import bl4ckscor3.bot.bl4ckb0t.commands.channel.Source;
 import bl4ckscor3.bot.bl4ckb0t.commands.channel.Stop;
 import bl4ckscor3.bot.bl4ckb0t.commands.channel.Twitch;
 import bl4ckscor3.bot.bl4ckb0t.commands.channel.Twitter;
+import bl4ckscor3.bot.bl4ckb0t.commands.channel.Update;
 import bl4ckscor3.bot.bl4ckb0t.commands.channel.Version;
 import bl4ckscor3.bot.bl4ckb0t.commands.channel.Vowels;
 import bl4ckscor3.bot.bl4ckb0t.commands.channel.Weather;
@@ -100,6 +101,7 @@ public class Listener extends ListenerAdapter<Bot>
 		commands.add(new Stop());
 		commands.add(new Twitch());
 		commands.add(new Twitter());
+		commands.add(new Update());
 		commands.add(new Version());
 		commands.add(new Vowels());
 		commands.add(new Weather());